+++ app/controllers/auth_controller.rb
@@ -1,0 +1,22 @@
+class AuthController < ApplicationController
+
+    helper_method :current_user,
+                  :logged_in?
+  
+    private
+  
+    def authenticate_user!
+      unless current_user
+        cookies[:start_page] = request.url
+        redirect_to login_path, alert: 'Are you a Guru? Verify your name and password, please!'
+      end
+    end
+  
+    def current_user
+      @current_user ||= User.find_by(id: session[:user_id]) if session[:user_id]
+    end
+  
+    def logged_in?
+      current_user.present?
+    end
+end
