+++ app/controllers/test_passages_controller.rb
@@ -1,5 +1,5 @@
-class TestPassagesController < ApplicationController
-    
+class TestPassagesController < AuthController
+  before_action :authenticate_user!  
   before_action :set_test_passage, only: %i[show update result]
 
   def show; end
