+++ app/helpers/auth_helper.rb
@@ -1,0 +1,2 @@
+module AuthHelper
+end
