+++ app/views/layouts/application.html.erb
@@ -11,7 +11,7 @@
 
   <body>
     <%= render 'shared/nav' %>
-    <%= flash_message %>
+    <%= flash_message(flash.keys.first) %>
     <%= yield %>
     <%= render 'shared/footer' %>
   </body>
